Major changes in data model due to duplicated creation of projects in each participant project. The mapping between both objects was totoally broken, but should be better/fine now. However, requires to exclude the mapping from Teilnehmer to Projects.
Failed tests due to unknown error in updating the Teilnehmer object to add a registered project: Within the REST API, this object is update successfully, meaning that a Project could be added to a Teilnehmer object, stored, retrieved back and verified that the Project object is still there. This procedure fails, however, in the test.

Possible fix: Remove entirely the fields angemeldeteProjekte and stornierungen from the Teilnehmer class. Would ease the modeling, but requires a bit more of communication as only the Project objects would then store what Teilnehmer objects are registered.